Updated `bcg` function to provide dynamic output.
When adding a gem to the global Bundle configuration, it is always the
last line in the configuration file. Knowing this, we can print the
last line instead of having hard coded output.

Should Bundler change this behavior in the future, this could make the
output confusing, though. For now, this improves maintenance and output
accuracy.