Fixed `gld` function to handle single commits.
When displaying branch details where only a single commit existed on a
branch, the output would be empty. This happened because there was
nothing for Git log to compare. For example, `git log <sha>..<sha>`
where `<sha>` is identical. Instead, it's better to show the single
commit instead which is what this fixes.

This also fixes a situation where the first commit of the branch was
accidentally dropped in the detailed output. Using `<sha>^` corrects
this.