Updated IRB Ruby/Rails prompts.
Added project name to Ruby prompt. Example:

    [2.5.1|example]>

Added Rails version to Rails prompt. Example:

    [2.5.1|5.2.0|example|development]>

For the Ruby prompt, the project name is obtained via the Git project
root path. Safely ignores errors if using IRB in a directory that is
not a Git repo.

Lastly, the prompt code was refactored to clean up prompt construction
duplication.