Fixed empty directory creation when installing or linking files.
By adding a `mkdir.command` file to the root of a directory, the
Dotfile scripts will ensure the parent folder of that file will be
created without installing/linking the file.

This solves two purposes:

- Ensures we can check in empty directories to Git as, by default, Git
  will not track empty directories.
- Ensures these empty directories are created when installing and/or
  linking this Dotfiles project on new/existing machines.

This will initially be used to support the `.config/git/template`
directory in order to be empty, by default, so Git repos are
initialized with the bare minimum of files.