Added Git branch facsimile function.
Speeds up duplicating an existing feature branch under a new name,
immediately switches to it, and creates an empty commit to denote the
end of that work.

Handy in situations where you don't *necessarly* want to chain branch
but do need to continue working on code based off of the previous
branch (with the intent you'll either rebase or throw away the
duplicated commits at some point in your development process).