Updated Git branch SHAS to calculate range for master branch
Provides more flexibility in calculating full range of SHAs when
toggling between `master` and feature branches.

When on the `master` branch, only the last commit is used for the
range. This ensures all commits from root of repository to current
commit are included. This ends up looking like this:

    git log <last commit SHA>

When on a feature branch, the range is built up using the first commit
made on the feature branch plus the last commit. Example:

    git log <start of feature branch SHA>..<last commit SHA>