Added Git push new branch detection
Enhances the original `git push` alias to detect if the branch being
pushed to hasn't been created yet. If the branch is new, it will be
created with proper upstream tracking. Otherwise, changes are pushed as
they normal.

This will allow the `gpob` function to be deleted in the future since
the new `gp` functionality takes care of this use case.