Removed IRB readline configuration
This setting was originally added via this commit (8563dae2ee44) during
the Ruby 2.7.0 upgrade because the Reline gem wasn't compatible with
Readline. Due to recent upgrades to Reline, this is no longer an issue
and can be safely removed in order to make use of IRB multi-line
editing and syntax highlighting.