Added Terraform XDG plugins cache
Necessary to cache Terraform plugins. This seems to be the only way in
which to allow locally built plugins to coexist with downloaded plugins
from the Terraform registry. This means all plugins across all projects
will be located in this central cache instead of isolated per Terraform
project. Each individual Terraform project, in turn, will now symbolic
link to this cache.