Updated environment update function to include company and notification
Necessary to ensure company projects are updated along with Open
Source projects. Conditional logic was added to gracefully skip
unknown paths for systems that might not have an OSS or company
directory structure. Finally, a notification was added upon script
completion since it can take a while to complete and is handy to have
when multitasking.