Removed curl export from Bash configuration
This is a partial revert of e209db266fe4 where issues were encountered
with the upgrade to OpenSSL 3.0.5. The actual fix was to temporarily
disable the use of this line:

    export HOMEBREW_FORCE_BREWED_CURL=1

This allowed both OpenSSL to be reinstalled, along with the rest of
Homebrew to be upgraded. It's unclear as to why this cropped up to
begin with but everything is stable again.