Updated RuboCop Git hooks to analyze staged and feature branch files
Necessary to improve the intelligence of files being fed to RuboCop
for analysis. Previously, all project files were being analyzed but
this allows fine grained control where only staged files (if making a
commit) or feature branch files (if pushing to a repository) are
changed. Any file outside these boundaries is ignored in order to
improve performance and narrow scope to only what has changed.

The running of RuboCop was extracted to a private function to further
reduce code duplication.