Updated Frum aliases to match Fast Node Manager pattern
Both of these programs are similar in design and nature so have opted
to update Frum to match Fast Node Manager alias names so that jumping
between Ruby and Node feels similar when managing versions. This
required renaming the Ruby Web Server function in order to not
conflict with the new Frum alias.