Fixed DNS cache flush behavior
The use of `killall mDNSResponderHelper` was introduced to improve robustness of clearing out the DNS cache especially when using a VPN client. Due to the code being harder read as a single line, the original alias was converted into a function of the same name for improved readability and maintenance.