Added IRB Reline color patch
Necessary to make autocomplete easier to read and work with. The monkey patch is an unfortunate side effect to make this work but hopefully can be removed once Reline allows for support of custom colors. Also, wow, the Reline source code is terrible. This would be so much easier to customize if dependency injection was properly used.