Added Git notes aliases
This is a partial revert of 87ad517fad3e on 2022-12-24 as I'm bringing back support for notes since they work well with tags. You definitely still don't want to use them on commits, especially with a rebase workflow, but tags are fine since the shas are meant to be permanent.

Milestone: minor