Updated IRB to use Reline face and type completion configurations
Necessary to customize the IRB method completion prompt. link:https://github.com/ruby/reline/blob/master/doc/reline/face.md[Documentation] and remove the original monkey patch for a cleaner implementation. This was made possible with the release of Reline 0.4.0.

Also picked up type completion which is new in IRB 1.9.0 and allows for chained type completion (previously not possible). This requires the Prism and RBS gems to be installed but works great.

Milestone: minor