Added JavaScript update all function
Necessary, much like the `rbua` function, to update all JavaScript project versions to the latest (or specific) version. This helps improve multi-project maintenance at the cost of being a nearly duplicate implementation of the original `rbua` function. Depending on how this evolves, this might need to be refactored further.

Milestone: minor