Added Git checkout alias
This is a partial restore of "Updated to Git 2.23.0 (7ad32d7c5a0d)" which introduced `git switch` for the first time while removing `git checkout` behavior. While the use of `git checkout` is vastly diminished with the introduction of `git switch`, there are still times when it's nice to checkout a specific commit which is why this alias is being restored.

Milestone: minor