Refactored attributes to global Git configuration root
Necessary to provide better global support as Git will default to looking at this path if no attributes are set for the local project. This is a revert of c5ad6eac1c36 (Fixed Git attributes template) which once wasn't allowed but is supported, by Git, once more.

Milestone: patch