Refactored Git aliases/functions to use parent branch and SHA functions
Necessary to simplify the logic now that we can use the newly added `_git_branch_parent` and `_git_sha` functions to calculate and obtain the parent branch of any branch you might be working on. This also fixes a bug where, when on the default branch (i.e. `main`), you'd have no commits to interact with.

Milestone: patch